backend/requirements.txt
flask==2.3.3
flask-cors==3.0.10
google-generativeai==0.3.0
SQLAlchemy==1.4.23
python-dotenv==0.19.0
langchain==0.0.300
python-jose==3.3.0
requests==2.26.0
pandas==1.3.3
